notify cardv2

+++ docker-start.sh
@@ -1,8 +1,28 @@
 #!/bin/bash
 
-docker compose down
-# Запуск для оновлення образу
-# docker compose build --no-cache 
+COMMAND=$1
 
-# Запуск контейнера
-docker-compose up --build
+# Функція для запуску
+run_compose() {
+  local build_flag=$1
+  echo "🚀 Запуск контейнерів..."
+  # --remove-orphans видалить контейнери, яких більше немає в конфігу
+  docker compose --env-file .env --env-file .env.local -f docker-compose.yml up $build_flag -d --remove-orphans
+  echo "✅ Готово! Контейнери запущені у фоні."
+  echo "Щоб переглянути логи, виконайте: docker compose logs -f"
+}
+
+if [ "$COMMAND" == "build" ]; then
+  echo "🛠 Перезбірка та запуск..."
+  run_compose "--build"
+elif [ "$COMMAND" == "stop" ]; then
+  echo "🛑 Зупинка сервісів..."
+  docker compose down
+else
+  echo "▶️ Звичайний запуск..."
+  run_compose ""
+fi
+
+#./docker-start.sh build — перезібрати та запустити.
+#./docker-start.sh — просто запустити те, що вже зібрано.
+#./docker-start.sh stop — зупинити все.
